refactor(aipanel): 拆出请求主链路/动作/自动审批 hook
- 新增 components/ai/useAIChatRequests.ts：发送消息（排队/临时会话/主题调色/@提及/斜杠展开）、followup 应答、重试/编辑/删除、上下文压缩与全量摘要子任务、会话恢复请求
- 新增 components/ai/useAIChatActions.ts：取消/停止并恢复/恢复任务、工具批准/拒绝/继续/终止、还原预览与应用、终端候选指派、队列冲刷
- 新增 components/ai/useAIAutoApprovalSettings.ts：任务级/全局级自动审批设置补丁与协同提示词预设更新
- AIPanel.tsx 3099 -> 1643 行
